pyproject.toml
[tool.poetry]
name = "sand-booking"
version = "0.1.0"
description = ""
authors = ["Tariq Jamal A <tariqjamal4267324@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.5"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.7.0"
whitenoise = "^6.9.0"
pytesseract = "^0.3.13"
pillow = "^11.3.0"
opencv-python = "^4.12.0.88"


[tool.poetry.group.dev.dependencies]
beautifulsoup4 = "^4.13.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
